Address comments & move setup to top level
