--- scheduler/api/grpc/doc.go
@@ -1,4 +1,0 @@
-/*
-See bazel/execution/... for further details on scheduler implementation of grpc remote exection & longrunning operation APIs
-*/
-package grpc
