--- scheduler/api/http/doc.go
@@ -1,4 +1,0 @@
-/*
-See common/endpoints/... for further details on scheduler implementation of http TwitterServer API
-*/
-package http
