+++ scheduler/client/README.md
@@ -1,7 +1,10 @@
-# Scoot API
+# Scoot Client
 
-Scoot API contains the Scoot Cloud API Thrift definition, Cloud Server code,
+Scoot Client contains the Scoot Cloud API Thrift definition, Cloud Server code,
 autogenerated thrift code in both Golang and Java, and client/CLI code for the API. 
+
+It consists of a cli directory that defines the command line interface for talking
+to scoot, and a scootcl directory with a main() that exercises that CLI. There is also a top level client.go file that issues the commands exercised by the CLI via thrift to the scheduler.
 
 ## Design and Summary
 The API itself is defined in scoot.thrift, as service CloudScoot.
@@ -24,7 +27,7 @@ The main elements here are:
 
 ##### Client
 
-The client code in scootapi supports creating a CLI client interface to the Cloud API.  The main interface is __CLIClient__, which defines an Exec function, which will execute a command specified on the command line against the Cloud API.
+The client code supports creating a CLI client interface to the Cloud API.  The main interface is __CLIClient__, which defines an Exec function, which will execute a command specified on the command line against the Cloud API.
 
 The __simpleCLIClient__ type implements this interface, and uses github.com/spf13/cobra for command line parsing. The client's communication to Cloud Scoot is done through a Dialer - see scoot/common/dialer.
