+++ scheduler/domain/README.md
@@ -8,5 +8,5 @@ The scheduler code contains the following packages:
 * __domain__ - scheduler go objects: jobs, tasks and states, and thrift versions of these objects
   * __gen-go/sched__ - generated code from the thrift definitions
   * __scheduler__ - interfaces and implementations for job scheduling
-  * __worker__ - interface for scheduler to run tasks on a worker
+  * __worker__ - interface for scheduler to starts tasks on a worker & manage responses re: status of in progress tasks.
     * __workers__ - implementations of the worker interface (polling worker, etc), that invoke a runner
